feat: integration test fix and improve
