+++ crate/server/src/routes/crypto/sign.rs
@@ -5,7 +5,9 @@ use actix_web::{
     web::{Data, Json},
 };
 use cosmian_kms_server_database::reexport::cosmian_kmip::kmip_2_1::{
-    kmip_operations::Sign as KmipSign, kmip_types::UniqueIdentifier,
+    KmipOperation,
+    kmip_operations::Sign as KmipSign,
+    kmip_types::{LinkType, UniqueIdentifier},
 };
 use cosmian_logger::trace;
 use serde_json::json;
@@ -15,7 +17,7 @@ use super::{
     CryptoApiError, CryptoResult, SignRequest, SignResponse as CryptoSignResponse, b64_decode,
     b64_encode, jose_to_kmip_params,
 };
-use crate::core::KMS;
+use crate::core::{KMS, retrieve_object_utils::retrieve_object_for_operation};
 
 /// `POST /v1/crypto/sign` — Detached JWS signature over arbitrary payload.
 ///
@@ -23,7 +25,7 @@ use crate::core::KMS;
 /// JWS Signing Input = `ASCII(BASE64URL(UTF8(JWS Protected Header)) || '.' || BASE64URL(JWS Payload))`
 ///
 /// The payload is NOT included in the response (detached content per RFC 7515 Appendix F).
-#[post("/v1/crypto/sign")]
+#[post("/sign")]
 pub(crate) async fn sign(
     req: HttpRequest,
     kms: Data<Arc<KMS>>,
@@ -34,9 +36,25 @@ pub(crate) async fn sign(
 
     trace!(user = user, "POST /v1/crypto/sign kid={}", body.kid);
 
+    // JWS convention (RFC 7515 §4.1.4): `kid` in the protected header identifies
+    // the public key that can verify the signature, not the signing private key.
+    // Look up the private key's linked public key; fall back to body.kid for
+    // symmetric keys and standalone keys that have no PublicKeyLink.
+    let signing_kid = {
+        let owm = retrieve_object_for_operation(
+            &body.kid,
+            KmipOperation::GetAttributes,
+            &kms,
+            &user,
+        )
+        .await
+        .map_err(CryptoApiError::from)?;
+        owm.attributes()
+            .get_link(LinkType::PublicKeyLink).map_or_else(|| body.kid.clone(), |l| l.to_string())
+    };
     let protected_header = json!({
         "alg": body.alg,
-        "kid": body.kid,
+        "kid": signing_kid,
     });
     let protected_json = protected_header.to_string();
     let protected_b64 = b64_encode(protected_json.as_bytes());
